@@ -5,14 +5,15 @@ buildscript {
     }
 
     dependencies {
-        classpath("com.jetbrains.rd:rd-gen:0.1831.17") // https://www.myget.org/feed/rd-snapshots/package/maven/com.jetbrains.rd/rd-gen
+        classpath("com.jetbrains.rd:rd-gen:2022.3.2") // https://search.maven.org/artifact/com.jetbrains.rd/rd-gen
     }
 }
 
 plugins {
     id 'java'
-    id 'org.jetbrains.kotlin.jvm' version '1.3.61'
-    id 'org.jetbrains.intellij' version '0.4.19'
+    id 'org.jetbrains.kotlin.jvm' version '1.7.20'
+    id 'org.jetbrains.intellij' version '1.9.0'     // https://github.com/JetBrains/gradle-intellij-plugin/releases
+    id 'me.filippov.gradle.jvm.wrapper' version '0.14.0'
 }
 
 repositories {
@@ -21,29 +22,24 @@ repositories {
 }
 
 instrumentCode {
-    compilerVersion = "213.6461.79" // latest build of com.jetbrains.intellij.java from https://www.jetbrains.com/intellij-repository/releases/
+    compilerVersion = "222.4345.14" // latest build of com.jetbrains.intellij.java from https://www.jetbrains.com/intellij-repository/releases/
 }
 
 group 'com.reallyliri.plugins'
-version '1.4.2'
-
-sourceCompatibility = 11
-targetCompatibility = 11
+version '1.4.3'
 
 compileKotlin {
-    kotlinOptions { jvmTarget = "11" }
+    kotlinOptions { jvmTarget = "17" }
 }
 
 dependencies {
-    testCompile group: 'junit', name: 'junit', version: '4.12'
 }
 
 intellij {
     // see https://www.jetbrains.com/intellij-repository/snapshots
     type = 'RD'
-    version = "2021.3-SNAPSHOT"
+    version = "2022.2-SNAPSHOT"
     downloadSources = true
-    intellij.updateSinceUntilBuild false
 }
 
 publishPlugin {
